backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.20
pydantic[email]==2.9.2
cryptography==43.0.1
python-dotenv==1.0.1
alembic==1.13.2
pytest==8.3.3

# Person A stack
google-genai==1.70.0
paddleocr==2.7.3
paddlepaddle==3.3.1
pdf2image==1.17.0
Pillow==10.4.0
